requirements.txt
pika
django-extensions
django-rest-framework
drf-spectacular[sidecar]
djangorestframework-simplejwt